requirements.txt
eyed3==0.9.6
mutagen==1.45.1
Pillow==8.4.0
pypresence==4.2.0
PyQt5==5.15.6
python_vlc==3.0.12118
youtube_dl==2021.1.3